Update README.md
I added in our use case. I know we reviewed some of the use cases, but we should also go over all the use cases. Then divide them into what will be accomplished this round and then have a future "wish list" of use cases that would be worked on in the future. We could do this with the group tomorrow, if time.